Make modifications to setup for running in parallel using PEST++ and YAMR.
I didn't really modify anything scientifically/mathematically for this run.
Mostly re-arranging the way files are setup, or are expected to be setup
for running a master process and many slaves.

Right now, the expectation is that you have ~/dvm-dos-tem and ~/pestdemo
and that you will end up with a bunch of directories in ~/ (one for each slave)
The paths set in all these scripts rely on this being the case...

tussock_full.pst
 - change to do up to 4 optimization iterations
 - change paths to be relative to current directory

setup-slaves.sh, cleanup-slaves.sh
 - script to setup 6 distinct folders for running in parallel. the first
   one will be the master, rest are slaves. also add script to clean up folders

start-slaves.sh
 - starts the pestpp slave process in 5 separate sub directories. redirect
   output to a file, also stderr, and then put process in the background

pest-helper.py
 - add option for specifying location of caljson files.

wrapper script:
 - change to absolute path for locating program
 - remove tweak to calibration_directives.txt (assume it will be copied from
   dvmdostem directory in the correct state)
 - change call to pest-helper.py to reference a relative path for where to
   look for json outputs.